+++ libc/string/memcmp.c
@@ -1,9 +1,9 @@
 #include "../include/string.h"
 
-int memcmp(const void* aptr, const void* bptr, size_t size) {
+int memcmp(const void* aptr, const void* bptr, uint16_t size) {
 	const unsigned char* a = (const unsigned char*) aptr;
 	const unsigned char* b = (const unsigned char*) bptr;
-	for (size_t i = 0; i < size; i++) {
+	for (uint16_t i = 0; i < size; i++) {
 		if (a[i] < b[i])
 			return -1;
 		else if (b[i] < a[i])
