+++ libc/string/strlen.c
@@ -1,7 +1,7 @@
 #include "../include/string.h"
 
-size_t strlen(const char* str) {
-	size_t len = 0;
+uint16_t strlen(const char* str) {
+	uint16_t len = 0;
 	while (str[len])
 		len++;
 	return len;
